REFACTOR: extract NoResult and GifImage to a separate composable function
